Backend/requirements.txt
flask
flask-session
spotipy
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
isodate
fastapi